requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==3.6.2
asyncio==3.4.3
attrs==23.2.0
BingImageCreator==0.5.0
Brotli==1.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
Markdown==3.6
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.5
prompt-toolkit==3.0.38
Pygments==2.14.0
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
sniffio==1.3.0
socksio==1.0.0
sydney-py==0.21.0
urllib3==1.26.15
wcwidth==0.2.6
websockets==12.0
yarl==1.9.4
